+++ src/main/kotlin/com/tezos/lang/michelson/editor/intention/WrapSimpleTypeIntention.kt
@@ -1,0 +1,28 @@
+package com.tezos.lang.michelson.editor.intention
+
+import com.intellij.codeInsight.intention.PsiElementBaseIntentionAction
+import com.intellij.openapi.editor.Editor
+import com.intellij.openapi.project.Project
+import com.intellij.psi.PsiElement
+import com.tezos.lang.michelson.psi.MichelsonPsiFactory
+import com.tezos.lang.michelson.psi.PsiSimpleType
+
+/**
+ * @author jansorg
+ */
+class WrapSimpleTypeIntention : PsiElementBaseIntentionAction() {
+    init {
+        text = "Wrap in parentheses"
+    }
+
+    override fun getFamilyName(): String = "Tezos/Michelson"
+
+    override fun isAvailable(project: Project, editor: Editor?, element: PsiElement): Boolean {
+        return element.parent is PsiSimpleType
+    }
+
+    override fun invoke(project: Project, editor: Editor?, element: PsiElement) {
+        val replacement = MichelsonPsiFactory.getInstance(project).createComplexType(element.parent.text)
+        element.parent.replace(replacement)
+    }
+}