+++ Changelog.md
@@ -1,5 +1,7 @@
 # Version 1.0.0
 - Feature: Smart completion of instructions. Only those instructions which are compatible with the current stack are suggested.
+- Feature: Intentions to wrap a type in parentheses, e.g. "int" to "(int)"
+- Feature: Intentions to unwrap a type, e.g. "(int)" to "int"
 - Don't suggest section names inside of sections or types
 
 # Version 0.10.0